requirements/dev.txt
-r base.txt
django_nose
